extract dragon fixtures into dragon conftest.py
