+++ conftest.py
@@ -1022,81 +1022,3 @@ def prepare_db(
 
         return PrepareDatabaseOutput(db, new_db)
     return _prepare_db
-
-
-class MsgPumpRequest(t.NamedTuple):
-    """Fields required for starting a simulated inference request producer."""
-
-    backbone_descriptor: str
-    """The descriptor to use when connecting the message pump to a 
-    backbone featurestore.
-    
-    Passed to the message pump as `--fs-descriptor`
-    """
-    work_queue_descriptor: str
-    """The descriptor to use for sending work from the pump to the worker manager.
-    
-    Passed to the message pump as `--dispatch-fli-descriptor`
-    """
-    callback_descriptor: str
-    """The descriptor the worker should use to returning results.
-    
-    Passed to the message pump as `--callback-descriptor`
-    """
-    iteration_index: int = 1
-    """If calling the message pump repeatedly, supply an iteration index to ensure
-    that logged messages appear unique instead of apparing to be duplicated logs.
-    
-    Passed to the message pump as `--parent-iteration`
-    """
-
-    def as_command(self) -> t.List[str]:
-        """Produce CLI arguments suitable for calling subprocess.Popen that
-        to execute the msg pump.
-
-        NOTE: does NOT include the `[sys.executable, msg_pump_path, ...]`
-        portion of the necessary parameters to Popen.
-
-        :returns: The arguments of the request formatted appropriately to
-        Popen the `<project_dir>/tests/dragon/utils/msg_pump.py`"""
-        return [
-            "--dispatch-fli-descriptor",
-            self.work_queue_descriptor,
-            "--fs-descriptor",
-            self.backbone_descriptor,
-            "--parent-iteration",
-            str(self.iteration_index),
-            "--callback-descriptor",
-            self.callback_descriptor,
-        ]
-
-
-@pytest.fixture(scope="session")
-def msg_pump_factory() -> t.Callable[[MsgPumpRequest], subprocess.Popen]:
-    """A pytest fixture used to create a mock event producer capable of
-    feeding asynchronous inference requests to tests requiring them.
-
-    :returns: A function that opens a subprocess running a mock message pump
-    """
-
-    def run_message_pump(request: MsgPumpRequest) -> subprocess.Popen:
-        """Invoke the message pump entry-point with the descriptors
-        from the request.
-
-        :param request: A request containing all parameters required to
-        invoke the message pump entrypoint
-        :returns: The Popen object for the subprocess that was started"""
-        # <smartsim_dir>/tests/dragon/utils/msg_pump.py
-        msg_pump_script = "tests/dragon/utils/msg_pump.py"
-        msg_pump_path = pathlib.Path(__file__).parent / msg_pump_script
-
-        cmd = [sys.executable, str(msg_pump_path.absolute()), *request.as_command()]
-
-        popen = subprocess.Popen(
-            args=cmd,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE,
-        )
-        return popen
-
-    return run_message_pump
