+++ tests/dragon/test_featurestore.py
@@ -36,15 +36,12 @@ import pytest
 
 dragon = pytest.importorskip("dragon")
 
-from smartsim._core.mli.comm.channel.dragon_fli import DragonFLIChannel
-from smartsim._core.mli.comm.channel.dragon_util import create_local
 from smartsim._core.mli.infrastructure.storage.backbone_feature_store import (
     BackboneFeatureStore,
 )
 from smartsim._core.mli.infrastructure.storage.backbone_feature_store import (
     time as bbtime,
 )
-from smartsim._core.mli.infrastructure.storage.dragon_util import create_ddict
 from smartsim.log import get_logger
 
 logger = get_logger(__name__)
@@ -61,39 +58,6 @@ if t.TYPE_CHECKING:
 
 # The tests in this file must run in a dragon environment
 pytestmark = pytest.mark.dragon
-
-
-@pytest.fixture(scope="module")
-def the_storage() -> t.Dict[str, str]:
-    """Fixture to instantiate a dragon distributed dictionary."""
-    return create_ddict(1, 2, 4 * 1024**2)
-
-
-@pytest.fixture(scope="module")
-def the_worker_channel() -> DragonFLIChannel:
-    """Fixture to create a valid descriptor for a worker channel
-    that can be attached to. Does not modify environment vars."""
-    channel_ = create_local()
-    fli_ = fli.FLInterface(main_ch=channel_, manager_ch=None)
-    comm_channel = DragonFLIChannel(fli_)
-    return comm_channel
-
-
-@pytest.fixture(scope="module")
-def the_backbone(
-    the_storage: t.Any, the_worker_channel: DragonFLIChannel
-) -> BackboneFeatureStore:
-    """Fixture to create a distributed dragon dictionary and wrap it
-    in a BackboneFeatureStore.
-
-    :param the_storage: The dragon storage engine to use
-    :param the_worker_channel: Pre-configured worker channel
-    """
-
-    backbone = BackboneFeatureStore(the_storage, allow_reserved_writes=True)
-    backbone[BackboneFeatureStore.MLI_WORKER_QUEUE] = the_worker_channel.descriptor
-
-    return backbone
 
 
 def test_backbone_wait_for_no_keys(
